install express and rename lib to src
