+++ package.json
@@ -5,6 +5,7 @@
   "author": "tiuuuupj",
   "description": "A simple helpful robot for your Company",
   "dependencies": {
+    "express": "^4.16.3",
     "hubot": "^3.1.1",
     "hubot-diagnostics": "0.0.2",
     "hubot-google-images": "^0.2.7",
