+++ src/removeHeader.js
